wl12xx: triggering FW assert to perform recovery
Driver based FW recovery was modified. It is now trying to trigger
FW assert and waiting for WATCHDOG interrupt before continuing with
recovery.

This is done to allow FW to perform cleanups and halt gracefully
before recovery is performed. This allows for safe reading of
on-demand fwlog, core_dump, etc...

FW assert triggering is done by writing 0xcafecafe to FW address 0x1c
and then triggering PHY_RX_HEADER FW interrupt which is its highest
priority interrupt.

There is a timeout of 100ms waiting for WATCHDOG interrupt.

In case timeout is reached or we get ELP or SDIO errors, we resume
recovery without reading fwlog, core_dump, etc beacuse we assume that
FW is not responding...

Signed-off-by: Yair Shapira <yair.shapira@ti.com>