wl12xx: check for required fw change while enable or disable role
In certain scenarios our current hooks for checking required
fw change aren't enough and we could be left in MR even though
we've gone back to a single interface. This is of course not
good as we prefer running with the SR FW which contains goodies
like BTCOEX.

An example scanerio is when running P2P GO + STA and then
removing the P2P GO.

Fix it by adding checks for fw change also while enable or
disable roles.

Signed-off-by: Eyal Shapira <eyal@wizery.com>