wl12xx: FW recovery trigger assert bug fix
The bug was in the partition mapping. It seems that .mem3 partition that
we used to write 'magic' 0xCAFECAFE to address 0x1C is actually in use (it
contains the struct part of the FW_STATUS). It is in use although the size
of the partition is set to 0

The fix was to save old partition, map a new partition with .mem part set to
address 0x1C and remap previous partition after writing 0xCAFECAFE.

Signed-off-by: Yair Shapira <yair.shapira@ti.com>